restructure feed actions to folder
