+++ src/actions/actionTypes.js
@@ -143,15 +143,3 @@ export const RECEIVE_REPLY_DISLIKE = 'RECEIVE_REPLY_DISLIKE'
 export const RECEIVE_GROUP_REPLY_SUCCESS = 'RECEIVE_GROUP_REPLY_SUCCESS'
 
 export const RECEIVE_POST_REPLY_SUCCESS = 'RECEIVE_POST_REPLY_SUCCESS'
-
-export const FETCH_FEED_REQUEST = 'FETCH_FEED_REQUEST'
-export const FETCH_FEED_SUCCESS = 'FETCH_FEED_SUCCESS'
-export const FETCH_FEED_FAILURE = 'FETCH_FEED_FAILURE'
-
-export const RECEIVE_FEED_REPLY_SUCCESS = 'RECEIVE_FEED_REPLY_SUCCESS'
-
-export const RECEIVE_FEED_LIKE_SUCCESS = 'RECEIVE_FEED_LIKE_SUCCESS'
-
-export const RECEIVE_FEED_DISLIKE_SUCCESS = 'RECEIVE_FEED_DISLIKE_SUCCESS'
-
-export const RECEIVE_FEED_POST_SUCCESS = 'RECEIVE_FEED_POST_SUCCESS'
